Remove license restriction
I haved decided that the previously included extra license restriction, though entirely fair and deserved, violates the spirit of my decision to open-source this. Although the individual excluded did deserve consequences for their prior behavior, I have come to feel that this open-source repo is not an appropriate venue for trying to inflict punitive measures, even when, more generally, such measures are properly and justly deserved for arrogant and vile personal misbehavior. This is my view at the time of this writing.